Change the logic to write and read using the proper json format and other possible formats

+++ src/main/java/patterns/example/service/file/FileOperator.java
@@ -52,9 +52,12 @@ public class FileOperator {
         return sb.toString();
     }
 
-    public void removeContent(String filePath) {
+    public void removeContentIfExist(String filePath) {
         try {
-            new FileWriter(filePath, false).close();
+            File file = new File(filePath);
+            if (file.exists() && !file.isDirectory()) {
+                new FileWriter(filePath, false).close();
+            }
         } catch (IOException e) {
             throw new RuntimeException(e);
         }
